removed all extra files, moved protein logic into its own file, and started to reorder strucuture
