--- backend/API/BodyFormats/RequestClasses.py
@@ -1,6 +1,0 @@
-from fastapi import UploadFile
-from pydantic import BaseModel
-from typing import List
-
-
-# general classes
