--- backend/API/BodyFormats/ResponseClasses.py
@@ -1,13 +1,0 @@
-from pydantic import BaseModel
-
-
-class ProteinInfo(BaseModel):
-    """
-    The content related to a protein for the
-    frontend system
-    """
-    name: str
-    molecularWeight: float
-    color: str
-    id_num: str
-    id_str: str
