--- backend/Electro1D/Electro1DMain.py
@@ -1,12 +1,0 @@
-"""
-main class for 1d Electro1D
-Note: GUI-related fragments have been removed from this class.
-@author Landon Heatly & Bader Alharbi
-"""
-
-
-class Electro1DMain:
-    serialVersionUID = 6488065550619657403
-
-    def __init__(self):
-        super().__init__("Electrophoresis")
