--- backend/Electro1D/Plot.py
@@ -1,1 +1,0 @@
-# @author Amr Mualla, working on Plot.py & PlotTest.py
